refacotr🔨 : separate extension files (#7)
