+++ AppStore/AppStore/UIColor+labelColor.swift
@@ -1,0 +1,15 @@
+//
+//  UIColor+labelColor.swift
+//  AppStore
+//
+//  Created by 리아 on 2022/06/07.
+//
+
+import UIKit
+
+extension UIColor {
+
+    static let blackLabel = UIColor(white: 0, alpha: 0.8)
+    static let grayLabel = UIColor(white: 0, alpha: 0.6)
+
+}
